+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [BREAKING] Removed the `TransactionScriptBuilder` and associated errors from the `rust-client` (#781).
 * [BREAKING] Renamed "hash" with "commitment" for block headers, note scripts and accounts (#788, #789).
 * Added recency validations for the client (#776).
+* Updated protobuf bindings generation to use `miden-node-proto-build` crate (#807).
 * [BREAKING] Changed parameter of `with_expected_output_notes` to accept any type of `OutputNote` (#806).
 
 ### Fixes